--- src/main/java/com/example/mindbridge/dto/ApiResponseDTO.java
@@ -1,17 +1,0 @@
-package com.example.mindbridge.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class ApiResponseDTO<T> {
-    private boolean success;
-    private String message;
-    private T data;
-
-    public static <T> ApiResponseDTO<T> success(String message, T data) {
-        return new ApiResponseDTO<>(true, message, data);
-    }
-}
