--- src/main/java/com/example/mindbridge/dto/AuthResponseDTO.java
@@ -1,12 +1,0 @@
-package com.example.mindbridge.dto;
-
-import lombok.Data;
-
-@Data
-public class AuthResponseDTO {
-    private String token;
-
-    public AuthResponseDTO(String token) {
-        this.token = token;
-    }
-}
