--- src/main/java/com/example/mindbridge/entity/enums/Role.java
@@ -1,5 +1,0 @@
-package com.example.mindbridge.entity.enums;
-
-public enum Role {
-    USER, ADMIN
-}
