+++ src/main/java/com/mindbridge/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.example.mindbridge.config;
+package com.mindbridge.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
