+++ src/main/java/com/mindbridge/dto/ApiResponseDto.java
@@ -1,0 +1,17 @@
+package com.mindbridge.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class ApiResponseDto<T> {
+    private boolean success;
+    private String message;
+    private T data;
+
+    public static <T> ApiResponseDto<T> success(String message, T data) {
+        return new ApiResponseDto<>(true, message, data);
+    }
+}
