+++ src/main/java/com/mindbridge/dto/AuthResponseDto.java
@@ -1,0 +1,12 @@
+package com.mindbridge.dto;
+
+import lombok.Data;
+
+@Data
+public class AuthResponseDto {
+    private String token;
+
+    public AuthResponseDto(String token) {
+        this.token = token;
+    }
+}
