+++ src/main/java/com/mindbridge/dto/FindIdRequestDto.java
@@ -1,11 +1,11 @@
-package com.example.mindbridge.dto;
+package com.mindbridge.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class FindIdRequestDTO {
+public class FindIdRequestDto {
     private String username;
     private String phoneNumber;
     private String code;
