+++ src/main/java/com/mindbridge/dto/FindPasswordRequestDto.java
@@ -1,11 +1,12 @@
-package com.example.mindbridge.dto;
+package com.mindbridge.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class VerifyCodeRequestDTO {
+public class FindPasswordRequestDto {
+    private String userid;
     private String phoneNumber;
     private String code;
 }
