+++ src/main/java/com/mindbridge/dto/ResetPasswordRequestDto.java
@@ -1,11 +1,11 @@
-package com.example.mindbridge.dto;
+package com.mindbridge.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class ResetPasswordRequestDTO {
+public class ResetPasswordRequestDto {
     private String userid;
     private String newPassword;
     private String confirmPassword;
