+++ src/main/java/com/mindbridge/entity/enums/Role.java
@@ -1,0 +1,5 @@
+package com.mindbridge.entity.enums;
+
+public enum Role {
+    USER, ADMIN
+}
