+++ src/main/java/com/mindbridge/repository/UserRepository.java
@@ -1,8 +1,7 @@
-package com.example.mindbridge.repository;
+package com.mindbridge.repository;
 
-import com.example.mindbridge.entity.UserEntity;
+import com.mindbridge.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
