+++ src/main/java/com/mindbridge/util/SmsService.java
@@ -1,4 +1,4 @@
-package com.example.mindbridge.util;
+package com.mindbridge.util;
 
 import org.springframework.stereotype.Service;
